Add cdk code and build process steps
On deploy, this setup should hopefully give me an auto-scaling group
with one EC2 instance which has the nix closure app artefact installed.
I can then manually connect with SSH and install nix and then the
closure, to get a working binary 🤞

(Amigo doesn’t have a role yet that installs nix, so it doesn’t seem
straightforward to get it to start an ubuntu instance that has nix
installed, and getting it to start a NixOS instance seems even
trickier. This will hopefully do for now!)