Move cabal2nix step out of main build
This step is adding about 35 seconds to the build, which is annoying.
Instead I’ve checked in the resulting file: it does mean we need a
linting check to make sure the nix file has been updated from the cabal
file, but I think this is better for the common case.

(Using a separate linting job should mean we’ll prevent merging the PR
but avoid blocking the Riff Raff artefact production.)