Add basic reader for capi’s Content type
As a starting point, this commit adds a basic reader for capi’s Content
type (the one defined in thrift in content-api-models) and wires it
together with a new server endpoint “read-capi” which can accept content
as json and return the output via a pandoc writer.

Currently, html in string fields is not parsed: a nice next step would
be to parse that using pandoc’s html reader to get a more complete
document (and be able to render the body properly).