Parse html fields with pandoc’s html reader
Now things are looking a bit more complete!

But it’d be good to read more data out of the Content type, and possibly
to do the transforming in a more generic way, e.g. a typeclass like a
simpler ToJSON that converts from a Content into an expanded Content,
parsing html and so on.